Final changes to call hgettext support.
* Corrected main/Main.hs to use macros instead of Strings;
* Altered function "__" on GUIUtils to call upon getText. It uses unsafePerformIO, so that had to be imported as well.

As of now, this commit breaks leksah. I came to the conclusion that we included "__" calls to some Strings that are being looked for in prefs files and cabal files - which, upon translation, are never to be found, producing an error. This commit, then, serves more of a progress tracking purpose. The next one will probably have the fixes.