Add basic cabal sandbox and GHCJS support
New Package -> Snadbox menu can be used to issue the cabal sandbox
commands.  Also cabal sandbox sources are included in the workspace
under the sandbox they have been added to (so you don't need to include
them explicitly.  They are also included in things like grep and hlint.

The GHCJS support is very basic (we need a better way to support
different launch types).  For now there is a JS button on the toolbar
and when it is active (and you are not in debug mode) run will
load the jsexe output of GHCJS in the output pane (or your web browser
if you have not built with webkit support in leksah).