Fast lazy refresh of workspace pane.
New workspace pane was very slow because it populated the whole tree
and if you had lots of packages (and big ones like ghcjs-dom) in your
workspace it took a long time to start leksah and froze for a second
or two after each configure.  This change makes it lazy so that only
the children of expanded nodes are populated (and even then it only
adds the first child until the child node itself is expanded).