Use header bar for dialogs (fixes OSX fullscreen)
In OS X resizeable dialog boxes were shown fullscreen and took time to
show up.  This was slow and clunky and there windowUnfullscreen did not
prevent it.  CHanging to use dialogs with header bars for buttons fixes
this.  It also looks better in dark mode on OS X (no bright title bar
on the dialog box).