Fix #386 some refs are in Errors pane
I think the cause was removeLogRefs caused the error pane to be emptied
and completely repopulated as each file started to be built (even if
nothing needed to be removed).  Now it should only remove the items
matching the file in question.

There might be other improvements to be made, but for me this made a
big difference.