New function showDialogAndGetResponse and refactor SourceBuffer
In SourceBuffer, every dialog except one was refactored to use the new function. The one remaining in-line dialog code saves the dialog to an MVar in the process of creating it

The new showDialogAndGetResponse allows the caller to

    - Pass options to the new' constructor
    - Handle unexpected responses, such as ResponseTypeDelete, and
    - Collect state information (i.e. which response was chosen)

This makes it the most general of the showDialog* functions. In particular, the inability to remember which option was chosen makes showDialogOptions unsuitable as a starting point for SourceBuffer's dialogs